move mutations; add firebase directory
